+++ NEWS.md
@@ -1,3 +1,5 @@
+* Record precipitation totals from the weather observations at 51 minutes past the hour (#27) 
+
 * Add 'GMT' timezone to weather$time_hour and update code to match current file
   format from Iowa State mesonet (@rmcd1024, #24)